fix: address narrowing conversions for large tensors

+++ paddle/phi/kernels/cpu/rrelu_grad_kernel.cc
@@ -28,10 +28,10 @@ void RReluGradKernel(const Context& dev_ctx,
   const T* n_ptr = noise.data<T>();
   const T* x_ptr = x.data<T>();
   const T* out_grad_ptr = out_grad.data<T>();
-  int numel = static_cast<int>(x.numel());
+  int64_t numel = x.numel();
   if (!x_grad) return;
 
-  int i = 0;
+  int64_t i = 0;
   T* x_grad_ptr = dev_ctx.template Alloc<T>(x_grad);
   for (i = 0; i < numel; i++) {
     x_grad_ptr[i] = x_ptr[i] > 0 ? out_grad_ptr[i] : n_ptr[i] * out_grad_ptr[i];
